Add NRefactory 5.5.1 source code.
For the moment, we need to use a customized Mono.Cecil in ILSpy (thread-safe branch + System.Runtime fix),
and build NRefactory against that.

Add script for building ICSharpCode.Decompiler NuGet package.
This package is built against the NuGet Mono.Cecil and NRefactory packages, not against the sourcecode included in this repository.